feat: add upload items file feature
Adds a new feature to the logistic good received form that allows users to upload multiple files.

The user can select multiple files to upload, and the files will be stored in the server.
The list of uploaded files will be displayed in the form, and the user can delete any file from the list.
The uploaded files will be attached to the good received document when the user submits the form.

This feature allows users to upload multiple files related to the good received document, such as invoices, packing slips, and inspection reports.