feat: add quotation status column
Adds a 'quotation_status' column to the Quotation listing table. This column displays the status of the quotation, which can be one of the following: Draft, Sent, Accepted, or Rejected.

This allows users to easily see the status of their quotations and take appropriate actions, such as sending a quotation to a customer or accepting a quotation from a supplier.